requirements.txt
fastapi
uvicorn
python-dotenv
requests
PyPDF2

